hype: use separate threads for assemble() and packet sending
Basically this splits in different threads the code that calls the Lua
assemble() function and the one that actually sends packet.

Packets are pushed to a queue, and then popped from the send thread, packed
and sent. Rate limiting is done from the send thread instead of beeing done
before calling the Lua function like now.