hype: do the packet creation and sending in the same thread
Turns out that putting the packet sending in its own thread wasn't a very
good idea and it broke scripts like banner.lua and http.lua.

Basically the packets generated by the hype.send() function in the recv()
callback were always added after the packets generated by the loop()
callback (which run much faster) adding a ton of latency and making the TCP
connections to time out.

Now the queue is flushed by the same thread that calls loop() *before* loop()
is called so that the loop() doesn't saturate the queue. This affects the
overall performances (packets are sent more slowly) and needs a rewrite of
the rate limiting code.