pktizr: merge loop and send threads once again
This makes it possible to use the packet queue for the packets generated
on the recv thread only, meaning that the queue will generally be emptier
now and consume less memory then before, when, due to rate limiting, it
could end up consuming the whole memory (this was somewhat mitigated by
rate limiting packet creation as well as dispatch, which was a horrible
hack).

Additionally the latency between packet creation and dispatch is reduced
thus reducing bufferbloat.

However this means serializing packet creation (via Lua scripts) and
dispatch, which makes the inner loop slower (with standard Lua there's
a 15kpps performance drop, less with luajit).