script: use pkt struct as user data instead of copying
This makes use of Lua user data to store a pointer to the pkt struct, and
implements __index and __newindex methods appropriately, instead of always
doing the C struct <-> Lua table translation.

We make use of Lua __gc method as well so that the pkt structs are only
freed once they go out of scope in the Lua script. However, in case the
packet is sent using the pkt.send() function, it may go out of scope before
being sent, so a very primitive reference counting mechanism is used to
make sure that the packet is freed at the proper time.

Note that this slightly changes the Lua API: constructor functions (IP(),
TCP(), ...) no longer take a table to initialize the data.

Also, some code has been cleaned-up.