pkt_cookie: replace siphash with pyrhash
pyrhash is Python's randomized hash algorithm based on Fowler-Noll-Vo hash
function with randomized prefix and suffix. It's much simpler and faster
than siphash24, but it is *not* a cryptographic hash (not that we care
about that anyway).

Note that all of this is meant for my personal amusement only: while the
new hash function is faster than the old one, it doesn't affect performance
all that much.